Added C++ GCJ Framework.
Solutions can now be written in C++ with convenience of high-level
framework as in Python. What is more, it allows for parallelization with
TBB, when compiled with GCJ_PARALLELIZE macro defined.

Ported also 2009/round0/C problem to the C++ framework to test and
demonstrate its usage on a real problem.